order acceptance added

+++ src/main/java/com/jolyvert/deliveryapp/service/OrderService.java
@@ -5,5 +5,5 @@ import com.jolyvert.deliveryapp.exception.OrderInfoException;
 import com.jolyvert.deliveryapp.model.OrderInfo;
 
 public interface OrderService {
-    public OrderInfo createOrder(int cartId) throws OrderInfoException, FoodCartException;
+    public OrderInfo createOrder(long cartId) throws OrderInfoException, FoodCartException;
 }
